requirements.txt
Flask==3.0.3
loguru==0.7.2
mmh3==4.1.0
multiset==3.0.1
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
tqdm==4.66.2
tree_sitter==0.20.4
